simplify API. fix docs, add example
